Customer address update

+++ backend/src/main/java/com/horizon/repository/CategoryRepository.java
@@ -1,9 +1,7 @@
 package com.horizon.repository;
 
+import com.horizon.model.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.horizon.model.Category;
-
-public interface CategoryRepository extends JpaRepository<Category,Integer>{
-	
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 }
